+++ CHANGELOG.md
@@ -47,6 +47,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * (shield) Add new operators: `<`, `>`, `<=`, `>=`, `==`, `!=` for comparing integers to each others.
 * (shield) Add support for `string` objects
 * (shield) Add `contains(elem, array)` builtin function for checking if `elem` is contained in the `array`
+* (shield) Add `+`, `-`, `*`, `/` math operators for basic integer arithmetic
 
 ### Features